Refactor : 중복 예외 Global로 옮김
