+++ src/main/java/com/flab/delivery/exception/MenuException.java
@@ -2,15 +2,9 @@ package com.flab.delivery.exception;
 
 import org.springframework.http.HttpStatus;
 
-public class MenuException extends RuntimeException {
-    HttpStatus httpStatus;
+public class MenuException extends AbstractException {
 
     public MenuException(String message, HttpStatus httpStatus) {
-        super(message);
-        this.httpStatus = httpStatus;
-    }
-
-    public HttpStatus getHttpStatus() {
-        return this.httpStatus;
+        super(message,httpStatus);
     }
 }
