+++ src/main/java/com/flab/delivery/exception/handler/UserExceptionHandler.java
@@ -17,13 +17,6 @@ import javax.servlet.http.HttpServletRequest;
 @Slf4j
 public class UserExceptionHandler {
 
-    @ExceptionHandler(MethodArgumentNotValidException.class)
-    public CommonResult<Void> handleMethodArgumentNotValidException(MethodArgumentNotValidException ex, HttpServletRequest request) {
-        String msg = ex.getBindingResult().getAllErrors().get(0).getDefaultMessage();
-        log.info("Http Method : {}, URI : {}, msg : {}", request.getMethod(), request.getRequestURI(), msg);
-        return getBadResponse(msg);
-    }
-
     @ExceptionHandler(SignUpException.class)
     public CommonResult<Void> handleSignUpException(SignUpException ex, HttpServletRequest request) {
         log.info("Http Method : {}  URI : {}, msg : {}", request.getMethod(), request.getRequestURI(), ex.getMessage());
